consolidate to one reducer
