+++ frontend/src/components/DocumentPreview.js
@@ -3,8 +3,7 @@ import { connect } from 'react-redux'
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap'
 import { getAccessToken, updateDocumentStatus, getDocumentsByUser } from '../utils/ApiWrapper'
 import { bindActionCreators } from 'redux'
-import { updateDocuments } from '../redux/modules/user'
-import { beginLoading, endLoading } from '../redux/modules/auth'
+import { updateDocuments, beginLoading, endLoading } from '../redux/modules/user'
 import Iframe from 'react-iframe'
 
 import 'box-ui-elements/dist/preview.css'
