+++ frontend/src/redux/modules/reducer.js
@@ -1,13 +1,11 @@
 import { connectRouter } from 'connected-react-router'
 import { combineReducers } from 'redux'
 import user from './user'
-import auth from './auth'
 
 const createRootReducer = history =>
   combineReducers({
     router: connectRouter(history),
-    user,
-    auth
+    user
   })
 
 export default createRootReducer
